Disable CSRF protection
This is done because GitHub Webhook did not work
due to "CSRF verification failed" error with code 403.
This is a temporary solution until I understand what actually has to be
done to handle the error while keeping the CSRF protection enabled.

The Webhook is intended to send a request to '/hooks/build' URL
and thus trigger a LilyPond build process.