Dont compile scss with collectstatic
On scores deployment there was an error with `pysassc` in producion
environment. When new scores are compiled, the app calls `make static`
and therefore `scss` rule is also being executed, which is not needed.
Moreover, `pysassc` executable could not be found outside of the
virtual environment, so full path to `pysassc` had to be provided in
Makefile on production server.
Now `make scss` should be called manually before `make static`.
Additional rule can be created as the first rule in Makefile, so
both `make scss` and `make static` are called with one `make` command.